requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# ZoeDepth 深度估计依赖
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0
